requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.4.1
billiard==3.6.4.0
boto3==1.20.41
botocore==1.23.41
celery==4.4.7
certifi==2021.10.8
charset-normalizer==2.0.10
Django==2.2.6
django-celery-beat==2.2.0
django-celery-results==1.2.1
django-cors-headers==3.2.1
django-filter==2.0.0
django-mysql==3.8.1
django-timezone-field==4.2.3
djangorestframework==3.8.1
dnspython==2.2.0
eventlet==0.33.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.3
jmespath==0.10.0
jsonschema==2.6.0
kombu==4.6.11
Mako==1.1.6
MarkupSafe==2.0.1
mysqlclient==1.4.4
prometheus-client==0.9.0
pyparsing==2.4.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2019.3
PyYAML==6.0
redis==3.5.3
redis-py-cluster==2.1.0
requests==2.27.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
ujson==4.1.0
urllib3==1.26.8
vine==1.3.0
Werkzeug==1.0.1
zipp==3.6.0
